Allow no-bundle for faster startup

+++ lib/live_load/browser/connection/playwright/decompressor.ex
@@ -15,20 +15,28 @@ defmodule LiveLoad.Browser.Connection.Playwright.Decompressor do
   end
 
   defp do_extract(version) when is_binary(version) do
-    archive = Application.app_dir(:live_load, ["priv", "playwright", version, "playwright_bundle.tar.gz"])
+    base_path = Application.app_dir(:live_load, ["priv", "playwright", version])
 
-    if not File.exists?(archive) do
-      raise RuntimeError, """
-      The Playwright bundle does not currently exist.
+    dest = Path.join(base_path, "bin")
+    cli_path = Path.join(dest, "driver/playwright-driver")
 
-      In order to install Playwright, run `mix live_load.install` to install the default version.
-      """
+    if File.exists?(cli_path) do
+      cli_path
+    else
+      archive = Path.join(base_path, "playwright_bundle.tar.gz")
+
+      if not File.exists?(archive) do
+        raise RuntimeError, """
+        The Playwright bundle does not currently exist.
+
+        In order to install Playwright, run `mix live_load.install` to install the default version.
+        """
+      end
+
+      File.rm_rf!(dest)
+      File.mkdir_p!(dest)
+      :ok = :erl_tar.extract(String.to_charlist(archive), [:compressed, {:cwd, String.to_charlist(dest)}])
+      cli_path
     end
-
-    dest = Application.app_dir(:live_load, ["priv", "playwright", version, "bin"])
-    File.rm_rf!(dest)
-    File.mkdir_p!(dest)
-    :ok = :erl_tar.extract(String.to_charlist(archive), [:compressed, {:cwd, String.to_charlist(dest)}])
-    Path.join(dest, "driver/playwright-driver")
   end
 end
